Added a 'palette' field to terrain_map, for saving same.
Changes which ensure that every terrain_map has a palette.
Currently this is a best-effort solution; there exists at
least one case (merging two maps such as terrain combat map
+ vehicle map, that is done by overlay_map() I believe;
terrain-blitting mechs could do this also, I think.)
where, due to terrain being blitted from a map with another
palette (said terrain not present in the main map palette),
terrain can exist on a map which is not present in the palette.

In such cases, I soft-fail when writing the MAP {},
using ? or ?? or the like for any not-in-palette
terrain and adding a comment in the generated GhulScript
which notes that this was done.  This will not be sufficient
for the needs of 'save game', but is still very useful for
a map editor, where human hands can fix up the palette afterwards.

As to how to implement the Right Thing (TM), this brings up
some design questions.  Once terrain palettes are globally
scoped entities, rather than being defined for each MAP {},
there is no truly compelling reason to have more than one
palette; or at least one palette per glyph-width.  On the
other hand, there is no truly compelling reason to rip out
multiple palette functionality right now.  In the near future
(0.3.x branch), we should decide what is best.  Multiple palettes
would require palette-merging of moderate sophistication.
Such code is easy enough to write, but the requirement that it
never fail or generate an invalid palette is rather taxing.

In any event, here is one more step towards a map editor in-game.