Fix for 1425025: abuse of party moded + save game.
Fixed this by addressing the order in which objects are saved in the current
place. Previously, they were saved by tile order, so in addition to what the
bug report noted you could get some other odd things. Now, they're saved in
turn list order, and the current object is always saved at the head of the
list. This ensures that on reload the object which saved the game begins the
turn order and the place then continues executing objects in the original
order.