Fix for [ 1425039 ] Solo mode + possession = lockup
The charm works by setting a being's current faction to the
caster's. Character::setCurrentFaction() already checks if the character is a
party member, I added another check to see if the character was in solo mode,
and if so to set the party mode to follow.

Also found a bug where unreadyAll() was calling
player_party->unrefInventoryObject(arms) after unready() already did, causing
some oddities in the player inventory.