Bugfix: [1451180] Shift key adds letters in conversation
The new isprintable(char c) function which replaced libc's isprint was declared to
take a char arg. Some values like 304 (the value for the left shift key on my system),
when downcast to 8 bits, fell into the 32-127 range and were declared
printable. Fixed this by changing to isprintable(int c).