Just committed a CVS change to allow a smaller UI. In common.h, uncomment the
//#define CONFIG_11X11_MAP

line, recompile, and the screen will be 736x400. I've only done some limited
testing but things look surprisingly good. There might be a line-of-sight bug
and the splash screen needs some tweaks. Need to test some more and figure out
how to do this automatically and dynamically so rebuilds aren't required. By
reducing the console width from 42 to 30 characters the screen could be
640x400, which will satisfy the 640x480 crowd at least.