Adding a facility for the script to store a hash of     "name" => integer values, for the purpose of declaring policy via script which otherwise might require hard-coded constants in common.h or similar.
The first use of this feature will be to replace the use of
the constant NAZGHUL_BASE_ACTION_POINTS, so that the Action Point
costs for various actions are entirely controlled by the script.
    (Movement and attack AP costs are already in game.scm arms.scm,
    but some half dozen actions require that the kernal know
    some AP cost to use.)

The API for kern_intvar is nearly identical to that
for cfg.c cfg.h, save that integers are stored.

A seperate hash is used, as the data stored is of
a per-session nature, rather than a global-for-all-games config nature.

This data is not currently saved, but a function
kern_intvar_save() is provided.