Fix for 1571076 - colorized console problems
Added a 'glyph' module that will associate colors with chars on a per-character
basis instead of using state in the ascii module. When a color formatting
string is processed the result is string of colorized glyphs. These can now be
printed to the screen module or directly to the ascii module.