Fixing bug 1808794 by adding a new feature: wilderness parties will engage in existing wilderness combats and attack towns.
Whenever the current place has a parent that is wilderness, it will
periodically (depending on map scale) give the objects in the wilderness a
chance to take a turn. This means that existing wilderness parties will "see"
the player in towns or combat maps and pursue as usual. When they are adjacent
they will join combat, which means they will deploy their dudes at the nearest
map edge and commence fighting.

If they are in a ship they will disembark first, so the ship map won't appear
when they join. As usual they won't attack from the ship unless they are
adjacent. I also disabled cannon bombardment from ships while the player is not
at the same scale as the ship.

NPC parties won't follow the player through portals. That means they won't
follow into most dungeons. That's a different mechanism so maybe I'll address
that later.